WIP - Refactored into a spring boot project.
