+++ src/test/java/com/rutkowski/eventreservationapi/EventReservationApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rutkowski.eventreservationapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EventReservationApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
