+++ src/test/java/com/rutkowski/eventreservationapi/SeatingChartTest.java
@@ -1,13 +1,12 @@
-package seatingcharttests;
+package com.rutkowski.eventreservationapi;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import com.rutkowski.eventreservationapi.service.SeatingChart;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import seatingchart.SeatingChart;
 
 class SeatingChartTest {
